// src/types/CNDIConfig.ts
export type CNDIProvider = "aws" | "azure" | "gcp" | "dev";

export type CNDIDistribution =
  | "eks"
  | "clusterless"
  | "gke"
  | "aks"
  | "microk8s"
  | "rke2";

export type CNDIVersion = "v2" | "v3";

export interface CNDIConfig {
  cndi_version: CNDIVersion; // default: v3
  project_name: string;
  region?: string;
  provider: CNDIProvider;
  distribution: CNDIDistribution;
  infrastructure: CNDIInfrastructure;
  applications: {
    [key: string]: CNDIApplicationSpec;
  };
  cluster_manifests: {
    [key: string]: unknown;
  };
}

/**
 * CNDIConfig["infrastructure"]
 */

export type CNDIInfrastructure = {
  terraform?: TFBlocks;
  cndi: {
    nodes?: Array<CNDINodeSpec>; // only omitted if distribution is "clusterless"
    network?: CNDINetworkConfig;
    functions?: CNDIFnsConfig;
    ingress?: CNDIIngressConfig;
    external_dns?: CNDIExternalDNSConfig;
    reloader: {
      enabled?: boolean; // default: true
      values?: Record<string, unknown>;
    };
    cert_manager?: {
      enabled?: boolean; // default: true
      values?: Record<string, unknown>;
      email: string;
      self_signed?: boolean;
    };
    microk8s: CNDIMicrok8sConfig;
    argocd: {
      hostname?: string; // auto ingress if set
    };
    open_ports?: Array<CNDIPort>;
  };
};

/**
 * CNDIConfig["applications"]
 */
export type ArgoAppInfo = Array<{ name: string; value: string }>;

export type IgnoreDifferencesEntry = {
  group: string;
  kind: string;
  managedFieldsManagers: string[];
  name: string;
  namespace: string;
  jqPathExpressions: string[];
  jsonPointers: string[];
};

export type ApplicationMeta = {
  name?: string;
  namespace?: string;
  labels?: Record<string, string>;
  finalizers?: string[];
};

export type ApplicationSyncPolicy = {
  automated: {
    prune: boolean;
    selfHeal: boolean;
  };
  syncOptions: string[];
  retry: {
    limit: number;
    backoff: {
      duration: string;
      factor: number;
      maxDuration: string;
    };
  };
};

export interface CNDIApplicationSpec {
  targetRevision: string;
  repoURL: string;
  destinationNamespace: string;
  chart?: string;
  path?: string;
  values: {
    [key: string]: unknown;
  };
  labels?: Record<string, string>;
  finalizers?: string[];
  directory?: {
    include?: string;
    exclude?: string;
  };
  info?: ArgoAppInfo;
  syncPolicy?: ApplicationSyncPolicy;
  metadata?: ApplicationMeta;
  ignoreDifferences?: Array<IgnoreDifferencesEntry>;
  revisionHistoryLimit?: number;
}

/**
 * CNDIConfig["infrastructure"]
 */

/**
 * CNDIConfig["infrastructure"]["terraform"]
 */
type TFBlockVariable = {
  type: "string" | "number" | "bool" | "list" | "map";
  description?: string;
  sensitive?: boolean;
};

type TFBlocks = {
  terraform?: {
    [key: string]: unknown;
  };
  provider?: {
    [key: string]: unknown;
  };
  variable?: {
    [key: string]: TFBlockVariable;
  };
  locals?: {
    [key: string]: unknown;
  };
  data?: {
    [key: string]: unknown;
  };
  resource?: {
    [key: string]: unknown;
  };
  module?: {
    [key: string]: unknown;
  };
  output?: {
    [key: string]: unknown;
  };
};
// ...
export type CNDINetworkConfig =
  | ({ mode: "create" } & CNDINetworkConfigCreate)
  | ({ mode: "insert" } & CNDINetworkConfigInsert);
// ...
export interface CNDINodeSpec {
  name: string;
  role?: NodeRole; // default: controller
  volume_size?: number;
  size?: number | string; // number: 500 or string: az machine type
  vm_size?: string; // az machine type
  disk_size_gb?: number;
  disk_size?: number;
  disk_type?: string;
  instance_type?: string;
  machine_type?: string;
  min_count?: number;
  max_count?: number;
  count?: number;
  labels?: Label;
  taints?: Taint[];
}

/**
 * CNDIConfig["infrastructure"]["cndi"]["functions"]
 */
export type CNDIFnsConfig = {
  hostname?: string;
  noModuleCache?: boolean;
  maxMemoryLimitMb?: number;
  cpuTimeHardLimitMs?: number;
  cpuTimeSoftLimitMs?: number;
  workerTimeoutMs?: number;
  edgeRuntimeImageTag?: string;
};

/**
 * CNDIConfig["infrastructure"]["cndi"]["microk8s"]
 */
type CNDIMicrok8sConfig = {
  addons: Array<Microk8sAddon>;
  version?: string; // 1.27
  channel?: string; // stable
  "cloud-init": {
    leader_before: Array<string>; //
    leader_after: Array<string>;
  };
};

type Microk8sAddon = {
  name: string;
  enabled?: boolean;
  args?: string[];
};

/**
 * CNDIConfig["infrastructure"]["cndi"]["ingress"]
 */
export type CNDIIngressConfig = {
  nginx: {
    public: {
      enabled?: boolean; // default: true
      values?: Record<string, unknown>;
    };
    private: {
      enabled?: boolean; // default: false
      values?: Record<string, unknown>;
    };
  };
};

/**
 * CNDIConfig["infrastructure"]["cndi"]["external_dns"]
 */

type CNDIExternalDNSConfig = {
  enabled?: boolean; // default: true
  provider: ExternalDNSProvider;
  domain_filters: Array<string>;
  values?: Record<string, unknown>;
};
// ...
interface CNDIPort {
  name: string;
  service?: string;
  namespace?: string;
  number: number;
  disable?: boolean;
  private?: boolean;
}